Add vgmap and vertex buffer size options
Script will now output partial vgmaps (complete vgmaps are a command-line option) and there is an option to remove unused semantics so that the output matches the GPU format (the game stores 18 semantics with a total stride of 160 bytes in the IT3, but only 12 semantics with a total stride of 88 bytes is sent to the GPU).  The trim function is used to facilitate compatibility of meshes with injection using 3DMigoto.